feat: Refactor monolithic app/main.py into separate modules (refs #107)
