+++ app/manifest.py
@@ -4,17 +4,18 @@ from __future__ import annotations
 
 from fastapi import APIRouter, status
 
-from mcp_observability.schemas import Manifest, Capabilities
-
 from app.initialize import SUPPORTED_VERSION
-from app.resource_store import list_resources
 from app.prompt_store import list_prompts
+from app.resource_store import list_resources
+from mcp_observability.schemas import Capabilities, Manifest
 
 # Attempt to collect tool names from FastMCP registry (best-effort)
 try:
     from app.mcp_server import mcp as _mcp
 
-    _TOOL_NAMES = [t.__name__ for t in _mcp._tools]  # pyright: ignore[reportPrivateUsage]
+    _TOOL_NAMES = [
+        t.__name__ for t in _mcp._tools
+    ]  # pyright: ignore[reportPrivateUsage]
 except Exception:  # pragma: no cover
     _TOOL_NAMES = []
 
@@ -33,4 +34,4 @@ async def manifest() -> Manifest:  # noqa: D401
         resources=[r.id for r in list_resources()],
         prompts=[p.id for p in list_prompts()],
         tools=_TOOL_NAMES or None,
-    ) 
+    )
