+++ app/prompts.py
@@ -2,11 +2,10 @@ from __future__ import annotations
 
 from typing import Dict, List
 
-from fastapi import APIRouter, HTTPException, status, Query
-
-from mcp_observability.schemas import Prompt
+from fastapi import APIRouter, HTTPException, Query, status
 
 from app.prompt_store import get_prompt, list_prompts, render_prompt
+from mcp_observability.schemas import Prompt
 
 router = APIRouter(tags=["mcp"])
 
@@ -34,4 +33,4 @@ async def prompt_render(prompt_id: str, args: Dict[str, str]) -> Dict[str, str]:
     except ValueError as exc:
         raise HTTPException(status_code=400, detail=str(exc)) from exc
 
-    return {"prompt": rendered} 
+    return {"prompt": rendered}
