+++ app/security.py
@@ -1,6 +1,6 @@
-from fastapi import Header, HTTPException, status
-
 import os
+
+from fastapi import Header, HTTPException, status
 
 
 async def verify_bearer_token(authorization: str | None = Header(default=None)) -> None:
@@ -32,4 +32,4 @@ async def verify_bearer_token(authorization: str | None = Header(default=None)) 
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
             detail="Unauthorized",
-        ) 
+        )
