+++ tests/test_logs.py
@@ -3,7 +3,8 @@ import asyncio
 import pytest
 from httpx import AsyncClient, ASGITransport
 
-from app.main import app, _fetch_error_logs
+from app.main import app
+from app.routers.logs import _fetch_error_logs
 
 
 class DummyResponse:
@@ -49,7 +50,7 @@ async def test_fetch_error_logs_success(monkeypatch: pytest.MonkeyPatch):
         }
     }
 
-    monkeypatch.setattr("app.main.httpx.AsyncClient", lambda *a, **k: DummyClient(json_data=fake_json))
+    monkeypatch.setattr("app.routers.logs.httpx.AsyncClient", lambda *a, **k: DummyClient(json_data=fake_json))
 
     lines = await _fetch_error_logs(10)
     assert lines == ["first error", "second error"]
@@ -68,7 +69,7 @@ async def test_logs_errors_endpoint(monkeypatch: pytest.MonkeyPatch):
     }
 
     monkeypatch.setenv("MCP_TOKEN", "testtoken")
-    monkeypatch.setattr("app.main.httpx.AsyncClient", lambda *a, **k: DummyClient(json_data=fake_json))
+    monkeypatch.setattr("app.routers.logs.httpx.AsyncClient", lambda *a, **k: DummyClient(json_data=fake_json))
 
     transport = ASGITransport(app=app, raise_app_exceptions=True)
     async with AsyncClient(transport=transport, base_url="http://test") as ac:
