+++ tests/test_traces.py
@@ -1,7 +1,8 @@
 import pytest
 from httpx import ASGITransport, AsyncClient
 
-from app.main import _fetch_trace_json, _fetch_trace_logs, app
+from app.main import app
+from app.routers.traces import _fetch_trace_json, _fetch_trace_logs
 
 
 class DummyResponse:
@@ -31,7 +32,7 @@ class DummyClient:
 async def test_fetch_trace_json(monkeypatch: pytest.MonkeyPatch):
     fake = {"data": {"traceID": "abc"}}
     monkeypatch.setattr(
-        "app.main.httpx.AsyncClient", lambda *a, **k: DummyClient(json_data=fake)
+        "app.routers.traces.httpx.AsyncClient", lambda *a, **k: DummyClient(json_data=fake)
     )
 
     result = await _fetch_trace_json("abc")
@@ -42,7 +43,7 @@ async def test_fetch_trace_json(monkeypatch: pytest.MonkeyPatch):
 async def test_fetch_trace_logs(monkeypatch: pytest.MonkeyPatch):
     fake = {"data": {"result": [{"values": [["1", "log1"], ["2", "log2"]]}]}}
     monkeypatch.setattr(
-        "app.main.httpx.AsyncClient", lambda *a, **k: DummyClient(json_data=fake)
+        "app.routers.traces.httpx.AsyncClient", lambda *a, **k: DummyClient(json_data=fake)
     )
 
     logs = await _fetch_trace_logs("abc", 10)
@@ -60,11 +61,11 @@ async def test_trace_endpoints(monkeypatch: pytest.MonkeyPatch):
     async def fake_json(trace_id):
         return fake_trace
 
-    async def fake_logs(trace_id, limit):
+    async def fake_logs_func(trace_id, limit):
         return ["l1"]
 
-    monkeypatch.setattr("app.main._fetch_trace_json", fake_json)
-    monkeypatch.setattr("app.main._fetch_trace_logs", fake_logs)
+    monkeypatch.setattr("app.routers.traces._fetch_trace_json", fake_json)
+    monkeypatch.setattr("app.routers.traces._fetch_trace_logs", fake_logs_func)
 
     transport = ASGITransport(app=app, raise_app_exceptions=True)
     async with AsyncClient(transport=transport, base_url="http://test") as ac:
@@ -72,6 +73,6 @@ async def test_trace_endpoints(monkeypatch: pytest.MonkeyPatch):
         assert trace.status_code == 200
         assert trace.json() == fake_trace
 
-        logs = await ac.get("/traces/abc/logs", headers={"Authorization": "Bearer tok"})
-        assert logs.status_code == 200
-        assert logs.json() == {"logs": ["l1"]}
+        logs_resp = await ac.get("/traces/abc/logs", headers={"Authorization": "Bearer tok"})
+        assert logs_resp.status_code == 200
+        assert logs_resp.json() == {"logs": ["l1"]}
